requirements.txt
open-clip-torch
torch
lancedb
fastapi
uvicorn[standard]
pillow
pyarrow
transformers
accelerate
